Fix parent hard deleting child entity

+++ EventListener/SoftDeleteListener.php
@@ -10,6 +10,7 @@ use Doctrine\ORM\Mapping\JoinColumn;
 use Doctrine\ORM\Mapping\JoinTable;
 use Evence\Bundle\SoftDeleteableExtensionBundle\Exception\OnSoftDeleteUnknownTypeException;
 use Evence\Bundle\SoftDeleteableExtensionBundle\Mapping\Annotation\onSoftDelete;
+use Gedmo\Mapping\ExtensionMetadataFactory;
 use Symfony\Component\DependencyInjection\ContainerAwareTrait;
 use Symfony\Component\PropertyAccess\PropertyAccess;
 
@@ -138,6 +139,13 @@ class SoftDeleteListener
                     }
 
                     if ($objects) {
+                        $factory = $em->getMetadataFactory();
+                        $cacheDriver = $factory->getCacheDriver();
+                        $cacheId = ExtensionMetadataFactory::getCacheId($namespace, 'Gedmo\SoftDeleteable');
+                        $softDelete = false;
+                        if (($config = $cacheDriver->fetch($cacheId)) !== false) {
+                            $softDelete = isset($config['softDeleteable']) && $config['softDeleteable'];
+                        }
                         foreach ($objects as $object) {
                             if (strtoupper($onDelete->type) === 'SET NULL') {
                                 $reflProp = $meta->getReflectionProperty($property->name);
@@ -151,7 +159,11 @@ class SoftDeleteListener
                                     $property->name => array($oldValue, null),
                                 ));
                             } elseif (strtoupper($onDelete->type) === 'CASCADE') {
-                                $em->remove($object);
+                                if ($softDelete) {
+                                    $this->softDeleteCascade($em, $config, $object);
+                                } else {
+                                    $em->remove($object);
+                                }
                             } else {
                                 throw new OnSoftDeleteUnknownTypeException($onDelete->type);
                             }
@@ -160,6 +172,30 @@ class SoftDeleteListener
                 }
             }
         }
+    }
+
+
+    protected function softDeleteCascade($em, $config, $object)
+    {
+        $meta = $em->getClassMetadata(get_class($object));
+        $reflProp = $meta->getReflectionProperty($config['fieldName']);
+        $oldValue = $reflProp->getValue($object);
+        if ($oldValue instanceof \Datetime) {
+            return;
+        }
+
+        //check next level
+        $args = new LifecycleEventArgs($object, $em);
+        $this->preSoftDelete($args);
+
+        $date = new \DateTime();
+        $reflProp->setValue($object, $date);
+
+        $uow = $em->getUnitOfWork();
+        $uow->propertyChanged($object, $config['fieldName'], $oldValue, $date);
+        $uow->scheduleExtraUpdate($object, array(
+            $config['fieldName'] => array($oldValue, $date),
+        ));
     }
 
     private function getPropertyByColumName(\ReflectionClass $entityReflection, $name){
@@ -176,6 +212,5 @@ class SoftDeleteListener
                return $p;
             }
         }
-
-    }
+     }
 }
